Add files via upload
These are files for the fourth game programming hw and it is a simple platformer game in which you move only forward.  It is able to interact with the platforms, but the game will be updated till the bugs have been dealt with.  It is a simple traversal game in which you need to get to the finish goal.  Your sprite needs to simply touch it.  Do not complain about the graphic quality as that will be worked on, and I hope it is judged fairly.