HW 5
It is a simple program of SAT collisions between several rectangles different scale and rotations.  It is a demo for a future game that makes use of SAT collisions.  It shows collisions horizontally for simplicity to show that the SAT collisions work.  The program will implement vertical collisions in the future, but it will be a while.